packages/agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
redis
web3
requests
pydantic